root: tided up homebrew and chocolatey packagers' implementations
It seems there were some dirty executions found for homebrew and
chocolatey packagers in implementations. Hence, we need to tidy
them up for quality assurances. Let's do this.

This patch tides up homebrew and chocolatey packagers' implementations
in root repository.

+++ srcRUST/.ci/_package-chocolatey_unix-any.sh
@@ -202,8 +202,12 @@ if (-not (Test-Path \"\${data_dir}\\\\bin\\\\${PROJECT_SKU}.exe\")) {
 
 # Install
 Write-Host \"assembling workspace for installation...\"
-Move-Item -Path \"\${data_dir}\\\\bin\" -Destination \"\${root_dir}\"
-Move-Item -Path \"\${data_dir}\\\\lib\" -Destination \"\${root_dir}\"
+if (Test-Path -PathType Container -Path \"\${data_dir}\\\\bin\") {
+        Move-Item -Path \"\${data_dir}\\\\bin\" -Destination \"\${root_dir}\"
+}
+if (Test-Path -PathType Container -Path \"\${data_dir}\\\\lib\") {
+        Move-Item -Path \"\${data_dir}\\\\lib\" -Destination \"\${root_dir}\"
+}
 Set-Location \"\$current_dir\"
 Remove-Item \$data_dir -Force -Recurse -ErrorAction SilentlyContinue
 "
